requirements.txt
PySide6>=6.5.0
python-docx>=0.8.11
pywin32>=306
PyMuPDF>=1.23.0
